structure: move files into standard locations
Ok this is real mess. Partly the mess is from not
putting things in the right place at the start, and
partly because the project outgrew its initial vision.

The main lesson here is that it's fine to put everything
in one file demonstration purposes. But if the demo
outgrows the original purpose, files and specs and whatnot
need to put into standard locations.