Update to work with Ruby 3.2.2
A lot of small changes necessary to get this working
again, as the Ruby version 2.6.5 won't even compile on
M2 mac. However, with some tinkering, the code in
./lib/arscope.rb actually runs.

The deckrb gem is woefully out of date, and Alex Chaffee
appears to have dropped off the internet. After a fair
bit of poking around, the `Marp` extension for VS Code
does a fair job out of the box for the slides. There
will need to be some fairly serious work to get the
slides presentable again, but I think it will more tedious
then difficult. I don't think I'm going to get much LLM
help on it though. Will try anyway.