Presentation cleanup
Marp is just what I need right now. The markdown is a
little different from what deckrb (deck.js) expects, and
these changes update the following:

1. All the code blocks are now rendering properly.
2. All line breaks are treated as "hard" and long lines
   joined as necessary.

Next up for the presentation will be some modicum of
styling and theming, but there are code changes which
need to be dealt with before then.